Fix bugs in reusing trained model for multiple tests
The common scenario of tranining a model and re-using it multiple times
for subsequent tests was not properly handled. Both runtime crashes and logical
errors (for sets of the same size) would arise when eploying the same model
multiple times. This has now been fixed for both the newest "gurls_test" and the
default "gurls" testing processes.

* Distance is recomputed in the rbf part "predkernel_traintest" --previous if
was seemingly not used or save computation time in any process.
* Dummy seq introduced in "gurls_test".
* Extra "if" case in the convertFormat function.
* Minor appearance and header info edits.